split sqlite logic using build flags
